+++ pkg/js/CHANGELOG.md
@@ -6,9 +6,9 @@
 
 Changed:
 [BREAKING]
-- `friendlySyntaxToApiSyntax` is now `transformer.transformDslToJSON`
+- `friendlySyntaxToApiSyntax` is now `transformer.transformDSLToJSON`
 - `apiSyntaxToFriendlySyntax` is now `transformer.transformJSONToDSL`
-- `checkDsl` is now `validator.validateDsl`
+- `checkDsl` is now `validator.validateDSL`
 - All non-transformer functions have been removed
 
 You can read the pre-v0.2.0 changelog [here](https://github.com/openfga/syntax-transformer/blob/main/CHANGELOG.md)